commit on master

+++ loops/train.py
@@ -136,7 +136,7 @@ def val_ep(cfg, val_dataloader, classifier, writer, epoch, best_epoch,
         writer.add_scalar('val/loss', ep_loss / i, epoch)
         log_avg_metrics(writer, metrics_val, 'val', epoch)
         #epoch_score = torch.tensor(metrics_val['acc']).mean().item()
-        epoch_score = torch.tensor(metrics_val['mse']).mean().item()
+        epoch_score = torch.tensor(metrics_val['mse']).float().mean().item()
         print('VALIDATION ACCURACY: %f' % epoch_score)
         print('\n\n')
